Split out the decoder as well
